+++ Calculator/Calculator/CalculatorItemQueue.swift
@@ -1,0 +1,34 @@
+//
+//  Calculator.swift
+//  Calculator
+//
+//  Created by 이호영 on 2021/11/09.
+//
+
+import Foundation
+
+struct CalculatorItemQueue {
+    var queueList: LinkedList = LinkedList()
+    
+    mutating func enqueue(item: CalcultorItem?) {
+        guard let item = item else {
+            return
+        }
+        queueList.append(data: item)
+    }
+    
+    mutating func dequeue() -> CalcultorItem? {
+        if queueList.isEmpty {
+            return nil
+        }
+        let removedItem = queueList.removeFrist()
+
+        return removedItem
+    }
+
+    mutating func clearQueue() {
+        queueList.clear()
+    }
+}
+
+
